# backend/openapi.yaml
openapi: 3.1.0
info:
  title: UITGo Backend API
  version: 0.4.0
  description: |
    Minimal trip & auth API powering the UITGo rider/driver apps.
    Realtime updates stream over WebSocket at `ws://{host}/v1/trips/{tripId}/ws`.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/trips:
    get:
      summary: List trips for current user
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [rider, driver]
            default: rider
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: page
          in: query
          required: false
          description: Optional 1-based page index (takes precedence over limit/offset)
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: false
          description: Page size when using the page parameter
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Trip page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripListResponse'
        '401':
          description: Unauthorized
    post:
      summary: Create a trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '401':
          description: Unauthorized
  /v1/trips/{id}:
    get:
      summary: Get trip details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Trip fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '404':
          description: Trip not found
    patch:
      summary: Update trip status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400':
          description: Invalid request
        '404':
          description: Trip not found
...
  /notifications:
    get:
      summary: List notifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Notification page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '401':
          description: Unauthorized
  /notifications/{id}/read:
    patch:
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '204':
          description: Marked as read
        '401':
          description: Unauthorized
        '404':
          description: Notification not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TripId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    NotificationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TripResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        riderId:
          type: string
        driverId:
          type: string
          nullable: true
        serviceId:
          type: string
        originText:
          type: string
        destText:
          type: string
        status:
          $ref: '#/components/schemas/TripStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLocation:
          $ref: '#/components/schemas/LocationUpdate'
          nullable: true
    TripListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TripResponse'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    CreateTripRequest:
      type: object
      required:
        - originText
        - destText
        - serviceId
      properties:
        originText:
          type: string
        destText:
          type: string
        serviceId:
          type: string
    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/TripStatus'
    TripStatus:
      type: string
      enum: [requested, accepted, arriving, in_ride, completed, cancelled]
    LocationUpdate:
      type: object
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
    Notification:
      type: object
      required:
        - id
        - userId
        - title
        - body
        - type
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
        tripId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
    NotificationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
...